packaging: bump to current node.js (#122)
* packaging: bump to current node.js

Node 14 is in maintenance LTS releases, the active releases are on the Node 16 release path.  Node seems to indicate that Node 16 should be preferred.  This updates the target to use node 16 rather than node 14.  This behavioural chance also now matches the behaviour of the VSCode extension when it resided in SourceKit-LSP.

* Add --target=node16

Co-authored-by: Adam Fowler <adamfowler71@gmail.com>