Let TestExplorer cancel build while debugging
Move build task out of `runHandler` and run it for both run and debug. remove preLaunchTask from debug config so it doesn't attempt to build the tests again.

Minor event change as well. When reacting to packageResolved events await for the resolve to finish before continuing. Otherwise it attempts to run multiple resolves at the same time.