toolchain: remove dependency on `DEVELOPER_DIR` on Windows
This environment variable is not stable and should be avoided going
forward.  The toolchain no longer relies on it and will be removed in a
future toolchain release.  The paths which are being computed relative
to these variables can be computed in alternative manners.  This should
be safe with older releases and newer releases of the toolchain on
Windows.