Swift Task queue (#317)
* Add a task queue for running swift tasks

This is to ensure we aren't running swift tasks on top of each other.

* Test runner uses task queue

* Move UI update to task queue

* Remove commented out code

* Add flag for don't run if already active

If flag is set check activeOperation when adding a new operation to the task queue.

This is to fix an issue where sourcekit-lsp disagrees with local resolve and resolves run repeatedly. It occurs when packages are in edited state. This is actually a bug in SwiftPM libs (fixed on 5.7 branch) but for the moment I need a workaround.