Ensure we don't add folders twice
It was possible to add a folder twice, via addWorkspaceFolders and  focusUri.
If a package takes a long time to load during initialisation, a focus event can occur prior to the package being fully loaded. At this point because the folder for that package isn't setup it will attempt to add the package again.