Toolchain path is one folder up from swift bin path
Previously it was two folders up and swift executables were assumed to be in a usr/bin folder relative to the toolchain folder. It looks like this cannot be assumed so the toolchain path is now one folder up instead.

I have also verified that sourcekit-lsp is happy with this as a valid toolchain path