Discover tests via LSP if available
If the LSP is available and supports the two requests for listing tests
then use those to discover tests in the workspace/document. If these
requests are unsupported or if they fail, fall back to the existing
methods.

This patch uses the existing code to leverage the LSP's workspace/tests
method and adds the textDocument/tests for listing tests within
documents as they change.

These requests produce a collection of TestClass structures, which are
hierarchical. These are passed to one of the methods in TestDiscovery.ts
which diffs against the existing tree and adds/removes tests. This is
similar to how it already worked, but now supports arbitrary levels of
nesting in the TestClass structure.

This patch also adds preliminary support for listing swift-testing tests
through the LSP methods, as well as through `swift test --list-tests`. It
does not support finding swift-testing tests through the document
symbols.

Another small improvement is if listing tests fails via `swift test
--list-tests` we'll attempt a `swift build --build-tests` before trying
again. If the user was on the 5.10 toolchain and no build had yet been
performed they would get an error.

Swift-testing tests are filtered out of the list of available tests
until code is added to run them. This will be added with #757.

- Adds tests for parsing the results of `swift test --list-tests`
- Adds tests for parsing tests via document symbols

This patch addresses both #761 and #762.