Fix missing test locations on toolchains 5.10 and below (#773)
If the LSP is unavailable `swift test list` will be used to enumerate
the tests. This method cannot attach a location to the TestClasses.

The code pre #767 took this in to account, but #767 introduced a
regression that would overwrite an existing item enumerated by document
symbols which had a location with one that didn't.