Add a CustomExecution for `swift` tasks
To give us more control of the swift process and allow us
to consume the output from swift, introduce our own
CustomExecution. This will set us up to address #750, #653
and #694.

The CustomExecution will load the asar bundled node-pty
module that ships with vscode. This way we don't need to
build for every platform and arch