Add integration tests for TestExplorer
Add some integration tests that exercise finding and running tests,
ensuring they complete with the expected outcome. It currently tests
both normal test execution as well as debugging the tests.

This patch adds a new profile used when running the tests locally via
the Extension Tests task. This ensures that the workspace is always in a
clean and consistent state.

It adds some caps to the docker containers so we can run the debugger
in tests.

It also cleans up the test logs, silencing some harmless warnings.