Switch to Build All task when listing tests fails
Use the buildAllTask instead of SwiftExecOperation to make long this
behaviour the same as other builds in the extension.

Added a task cache since provideTasks can be called multiple times and
the most recently returned task is held by VSCode. If we attempted
created the build all task and hooked up the event listeners but then
provideTasks was called again before the task was executed, the executed
task would not be the one we were listening to.

Fixes #808