Provide error when debugger isn't started (#853)
* Provide error when debugger isn't started

When debugging an XCTest that fails to build the debugger promise
rejects but did not provide an error. The undefined error is caught and
passed to `getErrorDescription` which failed to read stderr off of it.
This caused another error to be thrown and caught at the top level,
showing a dialog.

Throw an error string instead of undefined, and handle the undefined
error case in getErrorDescription.