Remove stale swiftc diagnostics
Issue: #750

If the user modifies source and SourceKit diagnostics differ, they may
want to cleanup all the old swiftc ones. This will be turned on by
default. The swiftc must match with a former SourceKit one, and once
SourceKit returns a new list of diagnostics, if that matched diagnostic
is no longer there, assume it is fix and remove the matching swiftc error

Tie this to a setting so the user can turn off.

Keep track of all current diagnostics regardless of source to allow for
this cleanup, and has the added benefit that we can update the list
when the diagnosticsCollection setting changes.