Allow writing to different sourcekit output channels
Issue: #833

IFF sourcekit logMessage notification includes a logName property, we'll
use a different output channel to not polute the langauge server's main
log. For a known list of logs we'll set them up on startup so we can
control if the log level and/or timestamp is included in the log message.
For the current known indexing log, for example, all messages will be
"Info" so we'll just add timestamp. If new logs are introduced, we'll
show log level and timestamp by default