Run tests with code coverage
To help identify poorly tested areas of the extension, run the tests
with code coverage. Coverage is reported in the terminal as text as well
as stored in the `/coverage` directory as HTML.

In CI results can be seen in the terminal at the end of the test run.
Locally when running tests in the docker containers HTML coverage
results will appear in the `/coverage` folder after the run completes.