Add Swift: Capture VSCode Swift Diagnostic Logs command (#830)
* Add Swift: Capture VSCode Swift Diagnostic Logs command

Adds a new command that users can use to help generate bug reports for
the extension itself. The Capture VSCode Swift Diagnostic Logs command
will create a new folder in a temporary directory that contains:

- Swift version and path information
- Extension logs
- Any diagnostics in the Problems pane

Users can copy the folder path or, if they're on macOS, open it in
Finder. From there they can zip and attach these files to a GitHub
issue. We'll want to update the New Issue Template with instructions on
how to capture and attach these logs.