Use @vscode/test-cli to run tests
Migrates to the new @vscode/test-cli package to download the appropriate
VS Code and run the tests. This streamlines the test setup code, and
should be a more reliable way to run the tests as VS Code evolves.

Hopefully this also helps with the occasional SIGSEGV errors that happen
in CI on startup.