Improve specificity of ignored files in coverage reports
Code coverage reports are filtered to omit the tests themselves, as well
as all library code checked out in `.build`. This is done with the
`--ignore-filename-regex` argument passed to llvm-cov's `export` command.

The current filter is defined as
`Tests|swift-testing|Testing|.build|Snippets|Plugins`, and filters out
all files if they contain any of those strings in their filename. This
can be too aggressive, and was filtering out valid files under test and
omitting them from reports. For instance the file
`SwiftTestingScanner.swift` is filtered out of the coverage report for
the sourcekit-lsp repository despite being a valid file to be covered.

Instead, filter with fully qualified paths for each test target's source
directory and the `.build` folder. This omits Tests and library code
while avoiding the problem outlined above.