Use correct xctest path when debugging on Windows
The windows debug config was delegating to the macOS debugging config,
which runs xctests through a prebuilt executable included in the
toolchain. On Windows, like Linux, an xctest executable is built during
`swift test`/`swift build --build-tests`. This should be used instead.

This patch brings the Windows config in line with what we do on Linux.

Issue: #970