Improve error reporting when SourceKit-LSP fails to start
If SourceKit-LSP fails to initialize, typically because of unrecognized
launch arguments, don't attempt to restart it and limit the number of
erorr dialogs to just one descriptive one. Clicking the "Go to Output"
button on this dialog takes you to the SourceKit-LSP Output panel which
now shows the SourceKit-LSP launch error.

Issue: #976