Associate test run output with individual tests
For both XCTests and swift-testing associate each line of the output
with the appropriate test. This lets users view just the output for a
given test via the test history pane, and also shows the output for the
given test when navigating through the test run issues via prev/next
issue buttons. Clicking a test issue in the test history pane highlights
the issue inline and then clicking the "Go to Previous Test Failure"
button will show complete output captured for the test.

A current limitation is this does not capture output from user `print()`
calls in the same order they'd appear in stdout. We do the best we can
to capture prints from stdout and inject them in to the test output, but
because we're mixing this with output generated from the JSON event
stream the exact position of print calls in the output is not
guaranteed.

Hopefully this can be improved in the future and stdout output from tests
can be emitted on the JSON event stream, where we could then associate it
with a specific test.

The existing output parsing tests are expanded and reworked to more
completely assert against the state of a parsed test.