Support a Unified Test Binary (#986)
With the merge of https://github.com/swiftlang/swift-package-manager/pull/7766
a `swift build --build-tests` produces a single binary that contains
both XCTests and swift-testing tests.

Update the build configs when debugging tests to use this new unified
test binary if it exists.

To determine if we're running a unified binary on macOS we can check for
the presence of the swiftpm-testing-helper, which is used to run the
swift-testing tests in the unified binary.

This utility isn't required on Windows and Linux as the test binary can
be invoked directly. To determine if we're running a unified binary we
check if the latest build produced a .swift-testing binary in the build
folder. If it doesn't exist, or if it does exist but its `mtime` is
older than the last build start time, we know its a unified binary. The
`mtime` check is required in case the user updates their toolchain to a
version that produces a unified binary but still has a .swift-testing
binary in their .build folder left over from an old build.

In the future we can remove these checks once we're confident users on
6.0 and main toolchains have moved on to versions that generate a
unified test binary. Tentatively I propose this to be around when
Swift 6.0 is finalized and released.

This change only affects debugging tests. All other testing code paths
use `swift test` which lets SwiftPM automatically handle the switch to
the new unified binary.