Create separate Build All Release task
We were naively transforming the arguments of the returned Build All
task to create one that would run a release version of Build All.
Unfortunately the `task.execution` remained the same and so while upon
inspection the task looked like it would run a release build it was not.

Create two separate versions of the Build All task, one for Debug and
one for Release builds, and use the release version when building tests
in release mode.