Use build description to check for unified testing binary
We were checking the `mtime` of the `<ProductName>.swift-testing` binary
to determine if it was produced by the latest build. The issue with this
is that when you run tests over and over without making any code changes
the binary is never rewritten because it is already up to date.

Instead of checking the `mtime` of the .swift-testing binary in the build
folder to determine if it was produced, use the build `description.json`
and check the builtTestProducts list to see if the legacy
`.swift-testing` binary was produced by the latest build.