REPL: introduce a new jupyter notebook based REPL
VSCode has introduced a native jupyter notebook API. Wire up a "kernel"
for Swift to provide a native REPL experience within VSCode. This is
primarily a PoC implementation that demonstrates how such an
implementation could work.

The primary limitation in the current implementation is the IO
processing in the LLDB REPL. We can apply a heuristic to capture the
error message (termination on an empty string), but do not have such a
heuristic for the output and currently only capture the first line of
the output, leaving the remaining output in the buffer. Improving the
stream handling would significantly improve the experience.

Co-authored-by: Jeremy Day <jeremy@thebrowser.company>