Add New Swift File command (#1018)
* Add New Swift File command

Add a command that creates a new Swift File. Add it as an option to the
File > New File picker. Add it as an option when right clicking a
folder in a Swift project context. Add a keybinding of Opt+S Opt+N (or
Alt+S Alt+N) to create a new Swift file.