Fix coverage report
Issue was the dist/extension.js file was what vscode-test was running as
the extension as it was what was specified as the entry point in the
package.json. So The globalWorkspaceContextPromise helper we use in a lot
of tests to make sure the extension is activated, returns an instance of
the WorkspaceContext. Any behaviour stimulated through that context is
missed from the coverage report, and also any breakpoint hits are ignored.
So for example the BackgroundCompilation instance that is created on
extension activation was left out of the coverage report.

To fix, only bundle to a single extension.js file when packaging a vsix.
When debugging or running tests we transpile all the files individually.
The dist/extension.js file is still used when activating, but sourcemap
now works so the coverage report is accurate.

Issue: #1021