Mark XCTests as passed without a second pass over the output (#1034)
The lastTestItem was not being set right away when a test succeeded
because tests were being marked as passed in a second pass of the
XCTest output after a run completed.

Going by the removed comment this second pass was added to disambiguate
passing tests with the same suite/test name across multiple targets.

However, the test target is present in the `testName` used to look up a
test's index. For instance, it may look like
`TestTargetA/Suite/testFoo` and so would not be identical to
`TestTargetB/Suite/testFoo`, and so we get the correct `testIndex` for
either test.