Route swift-testing output to /dev/stdout, CONOUT$
Currently we are merging two streams of information to produce
swift-testing test output, the JSON event stream written to a named pipe
and the stream of data from stdout. This captures both testing events
and user outputs produced by things like `print()` statements.

However, this approach interleaves the two sources in an arbitrary
order. Print statements produced during a test run are typically read
after events on the named pipe, which makes it difficult to tell what
prints belong to what test.

In order to enforce the correct order we can configure swift-testing to
write its events to /dev/stdout (or CONOUT$ on Windows). Swift-testing
aquires a lock to the `fd` it is writing events and output to, so the
order of events and prints is correctly enforced.

This also simplifies a lot of the code that worked with the JSON event
stream as we can now rely solely on the task's stdout/stderr output
without the need for named pipes.

When parsing stdout we try and parse a JSON event that matches the
swift-testing event schema, and if we fail we print the line directly.
If it is parseable, we process it as a swift-testing JSON event and omit
it from the test run output.