Address flakey swift-testing test
The `topLevelTestPassing()` test asserts that a print statement is
captured during a test. However because of #1054 it is possible that
this print is saved to the test run after the swift-testing test end
JSON event has been recieved. Until this issue is addressed, add a 0.1
second delay to ensure the print is captured before the test end event.