Use swift-testing stdout as test results log verbatim
Instead of reconstructing the swift-testing output from a combination of
stdout and the JSON event stream, simply pass stdout through directly
and use that as the test results log. Use the JSON event stream to drive
test pass/fail/skip/issue/etc behaviour.

This resolves a race condition between the JSON event stream and stdout
interleaving output in an incorrect order.

This approach has the drawback of no longer associating individual lines
of output derived from the JSON event stream with a specific test,
however this was offering limited benefit since user output wasn't
captured in this way. Once every line of output can be accounted for in
the JSON event stream we can add this functionality back in with user
output included.

Issue: #1054