Add command to run all tests in parallel (#1065)
* Add command to run all tests in parallel

The command palette has entries for Running all tests, debugging all
tests, and generating code coverage for all tests. However because the
parallel profile is added as a 'standard' profile the same as 'run all
tests' is it not automatically included in the command palette.

This patch adds a command to run all tests in parallel to the command
palette. If a user wants to assign a keyboard shortcut to this command
it should now be possible from the Keyboard Shortcuts view.

Note that the parallel run only parallelizes XCTest executions.
Swift-testing tests are run parallel by default, and can be run
serialized on a per suite basis, but this test run profile has no effect
on them.

Issue: #1064

* Add licence header

* Fixup newline in package.json