Add line terminator to XCTest filter regex
The regex to specify a specific XCTest to run did not specify that there
should be no more text after the test ID. As a result, attempting to run
a test like `testFoo` would also run a test in the same suite called
`testFooBar`.

Swift-testing tests do not have the same issue as their IDs include the
function parentheses and argument labels, and no text can appear after
the closing paren.

Issue: #1084