Revert using universal debug adapter name. (#1101)
The CodeLLDB extension expects the debug adapter to be called `lldb`. If
we are using it and we don't call `registerDebugAdapterTrackerFactory`
with `lldb` as the adapter name then we fail to get logs in the
debugSessionCallback when running tests. This prevents XCTests results
from being parsed, and hides the swift-testing outout.

Revert a portion of #1024 that always set the debug adapter name to
`swift-lldb` and use `lldb` when we're using CodeLLDB as the debug
adapter.

Issue #1100