Don't mark swift-testing test case results as runnable
When running a swift-testing test that takes arguments the results are
added to the Test Explorer showing their pass/fail status. At this time
these tests cannot be run with a specific argument so these tests should
not be marked as runnable. This prevents the "Run", "Debug" and
"Coverage" buttons from appearing on the test items in the explorer.

Issue: #1147