Improve Test Cancellation
There are two ways to cancel a running test, one via the Cancel button
at the top of the test explorer, and the other using the Cancel button
in the test run tree in the bottom right of the Test Results output
panel.

We were handling cancellation via the first method, but the one in the
test run tree uses a different cancellation token to indicate a test run
should be cancelled that is found on the constructed vscode.TestRun.

In our situation we want both of these to do the same thing, cancel the
build if it is running and stop the test run if it is running.

Create a CompositeCancellationToken that wraps both these child tokens
and performs cancellation if either of them request cancellation.

This approach also centralizes the cancellation logic to the
TestRunProxy, removing the need to pass the test explorer's token
around to the TestRunner methods.

This patch also improves cancellation when debugging, ensuring that if
you're doing a swift-testing + XCTest debugging run and you cancel
during the first test type the second one wont start.

Finally, print a "Test run cancelled" message in the test run output to
make it easier to see if a run was cancelled. This is helpful when
you're looking through the test results tree at old runs.

Issue: #1145