Associate XCTest suite start/stop with test item (#1152)
Capture the XCTest output associated with suites and assocaite it with
the test item that appears in the test run history. This makes it easy
to see the start/end time output of a suite by clicking on it.

In order to capture the test target for a suite we need to wait for the
first test that executes within that suite in order to get the target.
For whatever reason, XCTest does not print a suites target, only a
test's when it starts. The approach is if a suite starts, queue output
up until the first test run, then associate the last line of queued
output with the test target found on the started test. Lines before the
last line are not assocaited with a test (these are typically the top
level suite like "Selected Tests" or the test target name.

Issue: #1148