Attachments Support
Work in progress patch to support swift-testing attachments. Currently
prints the attachments out as a report at the end of a test run. In
order to leverage them we need to set the --attachment-path, which is
not available on the command line so we add a configuration json file.

There is still more work to do to allow the user to specify their own
attachment path on a per workspace basis.

This will currently only work with https://github.com/swiftlang/swift-testing/pull/770
as it isn't merged yet.