Simpler test activation/deactivation in tests (#1214)
Favour using the new `activateExtensionForSuite` and
`activateExtensionForTest` methods, which automatically deactivate the
extension when the suite or test completes.

These methods take optional setup/teardown methods that run on
suite/test setup/teardown. If the setup method returns a promise, this
is run automatically on teardown. This is useful for setup methods that
modify settings, as they can now return the promise directly from
`updateSettings` and any modified settings will be reverted
automatically.

This also prints any logs captured to the output channel if a suite/test
setup/teardown method fails with an error, which was being silently lost
before.