Add "auto" mode to sourcekit-lsp backgroundIndexing setting
Now that backgroundIndexing is no longer experimental it can be enabled
by default in Swift 6.1.

This patch evolves the `sourcekit-lsp.backgroundIndexing` setting to be
an enum with three possible values, `true`, `false` or `auto`.

The `true` and `false` options function as they did previously. The new
`auto` setting will enable background indexing on Swift >= 6.1. To
enable background indexing on Swift 6.0 the user must still set this
setting to `true`, as this feature is experimental in 6.0.