Fix location for diagnostics generated from macros via swiftc
The format of messages generated from macro errors/warnings is different
than those generated normally. The parsing regex would capture the
location for the related info diagnostic with a backtick. The location
for the parent regex would be "#macro expansion" which when clicked
would open an error window in a VS Code tab stating that the file
"#macro expansion" could not be found.

This patch tweaks two things:
1. Makes the regex for capturing the filepaths of warnings/errors
   account for leading trivia
2. If the path for a diagnostic is invalid, wait until we parse the
   related infromation and use the URI from there.