Warn when using `console` logging methods (#1241)
Add an ESLint warning when using methods like console.log in the
extension code. It is preferable to use the `SwiftOutputChannel` so that
these logs are captured by the extension diagnostics mechanism, and so
that they are all grouped together in the same place when debugging.

Sources under the tests directory do not have this restriction.