Test suites in nested extensions have incorrect parent in test explorer
Because swift-testing suites can be defined through nested extensions
the tests provided might not be directly parented to the test target.
For instance, the test target might be `Foo`, and one of the child
`testItems` in the document tests response might be `Foo.Bar/Baz`.
If we simply attach the `testItems` to the root test target then the
intermediate suite `Bar` will be dropped. To avoid this, syntheize
the intermediate children just like we synthesize the test target.