Notify build is preparing before progress starts
The build progress notification would only appear once the progress
values were shown in the output, i.e. [12/122]. Until progress values
appear, show a "Preparing <task>" notification. This only needs to apply
when the `showBuildStatus` value is `notification` or `progress`, since
`swiftStatus` already shows a message in the bar immediately.

This patch also only sets the `showBuildStatus` setting on a Task if it
is coming from a user defined task. Previously it was being set on
synthetic tasks which were being cached, so when the user updated the
`showBuildStatus` setting it wouldn't take effect until they restarted
the extension.

Issue: #1322