Remove Expand Macro Test (#1324)
Currently, it seems like the Expand Macros test has some issues that
makes it fail intermittently when run on GitHub actions. These issues
seem to be hard to reproduce and likely require simplifying the
dependencies the test depends on. Remove the test until there is a
better way to test this to avoid noise.

Issue: #1315