Show project dependencies hierarchically (#1339)
* Show project dependencies hierarchically

Currently the Package Dependencies view lists all dependencies in a flat
list. This list not only includes those explicitly defined in the
project's Package.swift, but also all the dependencies of dependencies.

As such its difficult to tell at a glance how a dependency is included.
It can also be confusing to see dependencies in the list that you did
not explicitly add.

Instead, show a top level list that is only the dependencies explicitly
defined in Package.swift. Expanding one shows any child dependencies of
the dependency. The files in the dependency's folder are still shown as
well, just after any child deps.